Move assets folder
