+++ pamplefront/assets/fakedata.ts
