+++ pamplefront/src/shared/hero/hero.component.html
@@ -1,6 +1,6 @@
 <img 
     id="hero" 
     class="img-fluid" 
-    src="../../assets/hero.png" 
+    src="assets/hero.png" 
     alt="man sitting on a couch with a laptop">
