+++ pamplefront/src/shared/herointern/herointern.component.html
@@ -1,7 +1,7 @@
 <img 
     id="herointern" 
     class="img-fluid" 
-    src="../../assets/hero_intern.png" 
+    src="assets/hero_intern.png" 
     alt="woman thinking"
 >
