+++ pamplefront/src/shared/picdoqcm/picdoqcm.component.html
@@ -1,7 +1,7 @@
 <img 
     id="picdoqcm" 
     class="img-fluid" 
-    src="../../assets/pic_qcm1.png" 
+    src="assets/pic_qcm1.png" 
     alt="Two people with a big blue question mark in between"
 >
