+++ pamplefront/src/shared/picdoqcm2/picdoqcm2.component.html
@@ -1,7 +1,7 @@
 <img 
     id="picdoqcm2" 
     class="img-fluid" 
-    src="../../assets/pic_qcm2.png" 
+    src="assets/pic_qcm2.png" 
     alt="Small cute rounded creatures playing with a question mark"
 >
